refactor(agent): unify sync attempt accounting
