+++ packages/agent/vitest.unit.config.ts
@@ -126,7 +126,10 @@ export default defineConfig({
       "test/sync-responder-protection.test.ts",
       "test/sync-on-connect-retry.test.ts",
       "test/sync-on-connect-churn.test.ts",
-      "test/rfc64-sync-on-connect-scheduling.test.ts",
+      "test/rfc64-sync-on-connect-recovery-plan-queue.test.ts",
+      "test/rfc64-sync-on-connect-peer-job-accounting.test.ts",
+      "test/rfc64-sync-on-connect-swm-lanes.test.ts",
+      "test/rfc64-sync-on-connect-selected-retry.test.ts",
       "test/sync-on-connect-peer-scheduler.test.ts",
       "test/system-context-graph-policy.test.ts",
       "test/sync-requester-bailout.test.ts",
